chore: refactor with proptest
